requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Distributed Training
ray[default]>=2.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
einops>=0.7.0
tqdm>=4.66.0
PyYAML>=6.0.0
matplotlib>=3.7.0
